+++ Sources/RichTextKit/_Deprecated/RichTextContext+Deprecated.swift
@@ -5,5 +5,11 @@ public extension RichTextContext {
 
     @available(*, deprecated, renamed: "highlightingStyle")
     var standardHighlightingStyle: RichTextHighlightingStyle { highlightingStyle }
+
+    @available(*, deprecated, renamed: "textAlignment")
+    var alignment: RichTextAlignment {
+        get { textAlignment }
+        set { textAlignment = newValue }
+    }
 }
 #endif